pulled libusb specific codes out of the hid_linux file
